refactor(merge-config): extract merge-config
