+++ src/get.js
@@ -1,10 +1,10 @@
 var request = require('request');
 var Promise = Promise || require('es6-promise').Promise;
-var objectAssign = require('object-assign');
 var isOk = require('is-ok');
 
 var HTTPError = require('./error').HTTPError;
 var buildDetailUrl = require('./build-detail-url');
+var mergeConfig = require('./merge-config');
 var defaultConfig = {
   headers: {
     'User-Agent': 'https://github.com/pandawing/node-chrome-web-store-item-property'
@@ -17,7 +17,7 @@ var defaultConfig = {
 
 function get (identifier, userConfig) {
   return new Promise(function (resolve, reject) {
-    var config = mergeConfig(buildDetailUrl(identifier), userConfig);
+    var config = mergeConfig(buildDetailUrl(identifier), defaultConfig, userConfig);
     request(config, function (error, response, body) {
       if (error) {
         reject(error);
@@ -30,23 +30,6 @@ function get (identifier, userConfig) {
       resolve(body);
     });
   });
-}
-
-function mergeConfig(url, userConfig) {
-  userConfig = userConfig || {};
-  var opts = objectAssign({}, defaultConfig, userConfig);
-  opts.url = url;
-  var headerCandidate = [{}, defaultConfig.headers];
-  if (userConfig.hasOwnProperty('headers')) {
-    headerCandidate.push(userConfig.headers);
-  }
-  opts.headers = objectAssign.apply(null, headerCandidate);
-  var qsCandidate = [{}, defaultConfig.qs];
-  if (userConfig.hasOwnProperty('qs')) {
-    qsCandidate.push(userConfig.qs);
-  }
-  opts.qs = objectAssign.apply(null, qsCandidate);
-  return opts;
 }
 
 module.exports = get;
